add Core project, add fields to entity, refactoring
