+++ DescriptiveGeometry/Core/DG.Core/Entities/BaseEntity.cs
@@ -1,0 +1,8 @@
+﻿namespace DG.Core.Entities;
+
+public abstract class BaseEntity
+{
+    public int Id { get; set; }
+    public DateTimeOffset DateCreated { get; set; }
+    public DateTimeOffset DateUpdated { get; set; }
+}
