+++ DescriptiveGeometry/Services/AuthorizationService/AuthorizationService.DAL/Entities/UserEntity.cs
@@ -1,10 +1,10 @@
 ﻿using AuthorizationService.DAL.Entities.Enums;
+using DG.Core.Entities;
 
 namespace AuthorizationService.DAL.Entities;
 
-public class UserEntity
+public class UserEntity : BaseEntity
 {
-    public int Id { get; set; }
     public string Name { get; set; }
     public string Email { get; set; }
     public string Password { get; set; }
